+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed an 'Undefined array key' error that occurred when saving an element, if any Neo blocks were in formerly valid places in the field for them to exist
+- Fixed a bug where Neo fields could be created without any block types at the top level
 
 ## 3.3.6 - 2022-09-01